Adds support for running and refactoring for Manifestations
This code is ported from dev-creatures but attempts to fit the new design with EF Core. 
Adds Player object to represent the players character
Player inherits from Manifestation and Actor, similar to client
Adds ManifestationManager to handle behaviors for packets related to manifestations
Updates the database (includes migrations) for RunState to save player characters run state. This is required so when the character is loaded, the previous state is restored.
Adds packets for client and server run toggle.