Support for setting default character appearance and logout
Ports over the dev-creatures work.
Database methods for loading character appearance data.
Sets the default item templates appearance data when creating a character (now when you login it's the base armor set and not a base body)
Basic support for not disconnecting when logging out in-game (packet handling) but no state saving on logout yet.