Adds support for running, character appearance, and EntityGUIDs (#69)
* Removes depecrated fallback property
To support single file publishing, this needed to be removed and cleaned up for .NET 5.

* Fix timestamp schema
MySQL was complaining during EF migrations that the CURRENT_TIMESTAMP was not valid for datetime columns

* Adds support for running and refactoring for Manifestations

This code is ported from dev-creatures but attempts to fit the new design with EF Core. 
Adds Player object to represent the players character
Player inherits from Manifestation and Actor, similar to client
Adds ManifestationManager to handle behaviors for packets related to manifestations
Updates the database (includes migrations) for RunState to save player characters run state. This is required so when the character is loaded, the previous state is restored.
Adds packets for client and server run toggle.

* Adds additional manifestation client packets

* Support for setting default character appearance and logout

Ports over the dev-creatures work.
Database methods for loading character appearance data.
Sets the default item templates appearance data when creating a character (now when you login it's the base armor set and not a base body)
Basic support for not disconnecting when logging out in-game (packet handling) but no state saving on logout yet.

* Fix missing comma

* Adds EntityManager and fixes GetEntityID implementation

Refactor to use ulong for entity IDs 
Pooling the entity values

* Refactor entities and how GUIDs are created

EntityGUID now has 1 byte for encoding the EntityType (255 values) and 56 bits for an ID. 

Also removes the Player/Character class for now since it was empty and now has Manifestation > Actor > Entity. 

There wasn't a need for an entity manager in these commits, so it's excluded now for brevity.

* Fix crash when entering world

The game reserves some entity IDs so it was crashing with ID = 1.

* Addresses requested changes

Remove starting value, not needed now that EntityType starts at 1

* Minor refactor of EntityGUID to address feedback

* Fix mask length and make constant